pyproject.toml
[project]
name = "video-transcriber"
version = "0.1.0"
description = "video-transcriber, buinding for FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.69.0",
    "httpx[socks]>=0.28.1",
    "loguru>=0.7.3",
    "openai>=2.30.0",
    "pydantic-settings>=2.13.1",
    "socksio>=1.0.0",
]
